+++ libsplinter/src/events/reactor/mod.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+mod igniter;
 mod reactor_shutdown_signaler;
 
 use std::sync::{
@@ -25,9 +26,10 @@ use crossbeam_channel::{bounded, RecvTimeoutError, Sender};
 use futures::Future;
 use tokio::runtime::Runtime;
 
-use crate::events::ws::{Context, Listen, ParseBytes, ShutdownHandle, WebSocketClient};
+use crate::events::ws::{Listen, ShutdownHandle};
 use crate::events::{ReactorError, WebSocketError};
 
+pub use igniter::Igniter;
 pub use reactor_shutdown_signaler::ReactorShutdownSignaler;
 
 /// Reactor
@@ -170,48 +172,6 @@ impl Reactor {
 impl std::default::Default for Reactor {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-/// The Igniter is a channel that allows for communication with a Reactor runtime
-#[derive(Clone)]
-pub struct Igniter {
-    sender: Sender<ReactorMessage>,
-    reactor_running: Arc<AtomicBool>,
-}
-
-impl Igniter {
-    pub fn start_ws<T: ParseBytes<T>>(
-        &self,
-        ws: &WebSocketClient<T>,
-    ) -> Result<(), WebSocketError> {
-        let context = Context::new(self.clone(), ws.clone());
-        self.sender
-            .send(ReactorMessage::StartWs(ws.listen(context)?))
-            .map_err(|err| {
-                WebSocketError::ListenError(format!("Failed to start ws {}: {}", ws.url(), err))
-            })
-    }
-
-    pub fn send(
-        &self,
-        req: Box<dyn Future<Item = (), Error = ()> + Send + 'static>,
-    ) -> Result<(), ReactorError> {
-        self.sender
-            .send(ReactorMessage::HttpRequest(req))
-            .map_err(|err| {
-                ReactorError::RequestSendError(format!("Failed to send request to reactor {}", err))
-            })
-    }
-
-    pub fn start_ws_with_listen(&self, listen: Listen) -> Result<(), WebSocketError> {
-        self.sender
-            .send(ReactorMessage::StartWs(listen))
-            .map_err(|err| WebSocketError::ListenError(format!("Failed to start ws {}", err)))
-    }
-
-    pub fn is_reactor_running(&self) -> bool {
-        self.reactor_running.load(Ordering::SeqCst)
     }
 }
